Updated Image-Based Visualization Components
Modified the false color component to detect the type of the analysis
and assign units and conversion factor.

Added remove to HDR2TIFF and HDR2GIF to make sure the file will be
updated.

Added pcond lines for HDR2TIF and HDR2GIF to set the exposure of HDR
images for human eye