Bugs fixed in Daysim Hourly results reader
I have never tested the components for a simple simulation with no
blinds after I modified them until today. They needed some modifications
to work for all the cases. It is fixed now.

I changed the default simulation type for image-based simulation to be
luminance analysis. Used to be illuminance.