Tolerance in Glazing Ratio List Fixed
Fixed a small issue that occurs when a list of glazing ratios is
connected and this list happens to divide up the cardinal directions
precisely where a surface is facing.  For example, you connect up a
glazing ratio list of 4 values for 4 cardinal directions but you have a
lot of surfaces facing due SE, right at the point where these cardinal
directions split.  Now, when this happens, the surface will default to
the closest direction in the counter-clockwise direction as opposed to
giving a random result (as before).