Modified to run annual radiation analysis
I noticed that setting output_units to 1 return all 0 results and not
the radiation values however assigning type 2 for each point using
sensor_file_unit works! I think this is a bug in Daysim. For now this
modifications should let the user run the annual analysis for both
daylighting and radiation.