Color Zones Based on EP Result Is HERE!!!!!!
Any EP result can now be hooked up to this component to create colored
meshes but I think that it looks best when you preview the zone breps
with the output colors.  The component includes several features to
recognize the EP data that is coming into and normalize it by the floor
area if it makes sense to.  Three inputs allow users to slice it and
interact with it as they wish!

Also, the Read EP Result component was updated with a couple of more
outputs that I became curious about as I colored the zones. Furthermore,
an option was added to normalize energy outputs by floor
area upon reading of the result file.