Read EP Result Now Accepts Any Name for Zones
Before, the component was not able to parse zone names that were not the
default names set by HB.  Now, the user can select any names for zones
and any names for surfaces and the component should work.  Also, the
Color zone component has had its output of "zoneNumbers" changed to
"zoneNames" and the component now outputs the names of the zones instead
of just numbers.