Split Read EP Result in Two
Given that the Read EP Result had blown up into the largest component in
Honeybee, I decided to split it into two: One for reading results
related to zones and one for reading results related to surfaces.  The
latter is much longer and deserves to be its own component anyway.
Also, In this commit, I have made it possible to import surface data
even if the zone name is not in the surface name.  Lastly, I have
coordinated the User Object icons between these result readers and the
IDF reader.