Added control on exporting interior walls
User can decide if they want interior walls to be exported to radiance.
This makes it possible to use an energy model for daylighting analysis
with no need to change the model.
On the other hand not every interior wall is an air wall. Right now the
concept of air wall hasn't been really implemented to Honeybee so it
should be fine however it should be modified later to only remove
airwalls and not all interior walls.
I think I can close this issue now:
https://github.com/mostaphaRoudsari/Honeybee/issues/103