Added setpoint and setback control
Added a function that takes an annual schedule and updates setpoints and
setbacks inside all the related schedules and returned an updated
version.  https://github.com/mostaphaRoudsari/Honeybee/issues/95

@chriswmackey this should be ready for you to study. I also noticed that
you are writing the node connections inside the component manually. I
assume it should be fine for a single zone but have you tried it for
multiple zones and validate it against other platforms? I think we
should do that if you haven't already or just leave it to EP to write
the connections.