Initial Commit of EP Window Shade Generator
The workflow is almost complete and I am just committing because I made
some changed to HB_HB that I do not want to lose.  The final part is
just checking the blind material in the WriteIDF component and writing
it into the IDF (the correct text is already formatted).  I will finish
this off in a few hours right before work.