ViewFactor Calculator Complete
Just finished a pretty solid component for generating test points in a
zone and calculating view factor to each of the zone surfaces from the
points.  The operation uses the mesh-Ray intersect and can be run in
parallel so that it is super fast.  Tomorrow, I will be polishing off a
component to take these view factors and produce maps of indoor MRT.