Put In Extra Checks
I decided to just bake and delete the Rhino geometry with the
SplitBuildingMasses component and this avoids ever having a zone brep
with faces all facing inward.  With this good, I also put a just into
the ViewFactor component to be sure that all generated points are inside
the zone.