Added an arbitrary testPt for annual glare analysis
For some unknown reason annual glare fails if the testPt file is empty.
To avoid the failure now I add a single point to the file. I need to
check the source code of Daysim tomorrow and see what is the deal with
testPts for annual glare. My understanding is that they should have no
relation!