Initial Commit
I found it useful to have a component that deconstruct zone into
HBSurfaces. One of the use cases is to create zones and add glazing by
Honeybee components, then deconstruct the zone, replace the roof with a
customized roof and use createHBZone to create a new zone. I should
double check and make sure that create zone doesn't remove the initial
inputs