Updated Descriptions + Added Glass Matl
I added a glass material as it seems like a lot of people want it and
the air gap component is lonely without it.  Also, I re-vamped the
descriptions for all of the material components since many of them did
not have any.  Lastly, I put in some checks to make sure that people do
not put in crazy values for some parameters.