Importing Curved Surface Results Now Works
It is now possible to import curved surface results correctly and
visualize them with the other HB components.  There is an assumption
being made that each curved surface has only one window but I realized
that we don't even have a naming convention for that case yet.  Maybe we
could do that in the future.