Fixed the bug with surface types
The wrong surface types was because of two different reasons.
1. In Honeybee I was assigning some specific construction types to
surfaces (e.g underground wall) which wouldn't match with OpenStudio
types and as a result types were considered as walls/unknown and
OpenStudio was re-evaluating the types based on normal.

2. I wasn't clear that OpenStudio explicitly is using RoofCeiling. I
added a catch to assign RoofCeiling for roofs and ceilings. Now the
types are coming in fine and OpenStudio fixes the direction of the
normal so it looks to be all fine.

I still need to check normal direction/point order. Looks like it has
issues with some of the non-convex horizontal surfaces, as well as
random vertical surfaces.

@chiensiTB I made sure to use the latest version but I assume you have a
newer version locally. Please make sure to double check when pushing the
newer version to github.