Implemented Basic Zone Air Mixing Objects
I implemented an option to mix the air between zones on the Solve
Adjacencies component.  Right now, this is just a basic implementation
that assigns a flow rate for air mixing based on the amount of surface
area between adjacent zones and a factor that I pulled from experiments
that I had done with DesignBuilder.
In the future, I hope to give people the option to set the actual flow
rate and a schedule for the flow as they desire.
Also, while this capability technically is granting the capability of
turning adjacent surfaces into an air wall surface type, I do not yet
change them to this.  It seemed that Mostapha had some idea for how to
do this previously and I should ask him before doing so.
-Chris