Ideal Air Systems Overhaul
At the request of @mostaphaRoudsari and the realization that the
original Ideal Air HVAC template was good, I have returned the HVAC
system to it's original state with one exception:  the old template was
not linked to the outdoor air control object.  I took this as an error
and linked the system to the outdoor air object.  I also exposed the
ability to change a lot of the parameters including the HVAC
availability schedule and many of the aspects of the Ideal Air System
like heat recovery, air economizers, and demand controlled ventilation.