Temporary fix for systems 2 and 7
@chriswmackey this should address your need for systems 2 and 7 so you
can go ahead and work on results reader. @chiensiTB please check the new
component I had to add extra checks for cases that HVACDetails and
PlantDetails are None. I also added planDetails to HBZone.HVACSystems.
Would be awesome to have a json file that sets default values for
systems so once details is set to None component assigns them as inputs.
I can go ahead and extract them from your component but it is probably
easier for you to do it. I also noticed that there is an issue with
default glazing construction that I will try to take care of it on my
way to Philadelphia.