Update Honeybee_ExportEPC.py
This change translates Honeybee geometry to the EPC. It checks to make sure the Excel file is open and then makes the necessary changes. Excel remains open. Inputs are _HBZones, filename, and filepath. Output is EUI (Energy Use Intensity) in kWhr/m^2/yr Known issue: code does not recognize a read-only file and opens a new instance of Excel.