Import OspenStudio Results Implemented
The OpenStudio Results for systems 2 and 7 are now successfully
implemented.  I also put separate outputs on the Export to OpenStudio
component for ChienSi's SQL file and the meter outputs.  Before the
release, I would like to implement some of the stuff that I put in for
the IDF component on the OpenStudio component including:
1) Blinds and zone shades
2) Natural Ventilation
3) Air walls and Zone Mixing
4) Internal Mass
Hopefully I can get to this in time.