Finalized Output for cMap
I finally now have all of the E+ outputs that I need to accurately
estimate air temperature stratification, including the total internal
gain of each zone (light+equip+people+solar+heatingSys) and the total
airflow into the zone from the outdoors (natVent+infiltration).  I also
bring in the energy lost/gained through natural ventilation, now making
it possible to construct a sankey diagram of a building's energy flow (I
will do this later).

I decided to do away with 3 outputs that I could see as misleading or
are obsolete.  I took out the total electric energy since I think a lot
of people might see this as heating/cooling electric energy, especially
if they use it with VAV systems, and it was not (it was also an internal
calculation that was time intensive and easy to break).  I also took out
the separate zone beam and diffuse gains since, if a user really wants
to understand this distinction, they can do so on a surface-by-surface
basis.  Now, only the total solar gain is imported.