Initial commit for Pollination components
These two components prepare a pollinator and open pollination page. In
this version user can load the file from their system and then play with
the graph. The graph is very basic for now and needs good amount of work
to be really useful but it is still better than nothing! #279 in
@chriswmackey 's style POLLINATION IS HERE!!!! ;)