Added grid and fixed the colors
Almost ready for release. I need to add type check so users don't load
the wrong file. Also it is cool to have something when the page loads
initially instead of white page with nothing! Also I'm not happy with
font size and the fact that grid size for table is fixed and won't work
for many cases.