THE ERA OF THERMAL AUTONOMY HAS COME!!!
At last! I can say after a year-long development process that all of the
known bugs in the comfort maps are fixed and all of the features needed
to meaningfully interpret the results are in place.

This final commit is the keystone in the arch of the comfort maps.  It
allows users to calculate thermal autonomy (or the percentage of
occupied hours when a space is thermally comfortable using passive
means).  With the ability to calculate thermal autonomy from comfort map
results, we now not only have a metric for evaluating how passively
comfortable a space is (which can shift the thermal debate away from
HVAC gadgets and towards good passive design) but we now also have a
means of accurately calculating this metric over large rooms and
spatially visualizing it.

This commit represents the first time ever that anyone has able to
calculate thermal autonomy at such a high spatial resolution.  At the
time of this post, this capability exists only in Honeybee and in no
other software known to the author.

May the open source nature of Honeybee ensure that this high-accuracy
method of calculating Thermal Autonomy always remain free and open and
may we  see the dawning of a new era of discussion of our thermal
environment.