All Clear for EP 8.3 Support
I ran about 20 different test cases through EP version 8.3 and all of
them worked without errors or warnings.  I tried to make sure that the
cases were related to things I saw in the 8.3 change log   As a result,
I think that we are all clear for version 8.3 support.
In the process of running the test cases, I also discovered a bug in my
"Add Internal Mass to Zone" component.  I also made a small change to
the way the version is written into the IDF so that users don't get a
meaningless and annoying warning when they open the IDF in the
IDFEditor.