Added Ability to Split Windows Vertically
At last, I think I can say that this component does all of the major
cases that people are concerned about.  This last one was critical as
there seems to be a general consensus that splitting a window vertically
provides the best daylight while minimizing unwanted conduction through
a lot of glazing.  Also, if it is done right, it is ideal for
encouraging bouyancy driven ventilation.