Added run in background
It is mainly useful for parametric runs. If you set runIt to 2 or any
number larger than 1 the study will be ran in background. It's helpful
so you can keep working, but it can be confusing since you won't see
anything while it's running. I needed this to be able to work while
having studies running on my machine.