Fixed The Last of the Bugs With Energy Balance
... for models with simple ideal air systems, that is.

This one took a while to figure out and this was not the first time
that, amidst the huge number of outputs from E+, they don't have a
really important one.  This time, the missing output was the energy
added/removed to the ideal air stream by the outdoor air.

Now, I calculate this value from the air added/removed to the zone and
the heating/cooling of the ideal air "coil."  Finally, everything seems
to be balancing out for any model that I throw at it!