THERM Export Workflow Compete!!
The workflow to export geometry from Rhino to THERM is now complete.
Both boundary conditions and materials are coming in ok such that all
the user has to do is hit "simulate."  Now, I am just going to make a
component to import the results of the simulation back into GH and then
I will put an example file together.