Fixed naming for glazing surfaces
@chriswmackey now you can get the base glazing name by
glazingName.upper().split("_GLZP_")[0] in all cases.
I updated read Read EP Surface Results component and it worked for both
energyplus and openstudio runs for rectangular windows, and well as
non-rectangular and nonplanar surfaces. It would be great if you can
also give it a try and let me know if it works fine.