Added LBNL WINDOW Construction Importer
It allows you to easily assign constructions that you create in LBNL
WINDOW as window constructions for EnergyPlus.  The assigning of
constructions works for both E+ and OpenStudio but the frame properties
currently only work for E+ as it will currently take a lot more code to
implement.  I am hoping that WINDOW allows an export in an Openstudio
format at some point, which would make this a whole lot easier.