Initial Commit of HVAC System Cleanup
@mostaphaRoudsari , @LelandCurtis ,
This is not a fully working commit but I will put up another commit
tomorrow that is fully functional.  This is just making the structural
change from having a GH component for each and every OpenStudio HVAC
component to one where we assign general HVAC features with 3 basic
components: airDetail, heatingDetail, and coolingDetail.  The new
structure mimics OpenStudio's general way of setting air loops, hot
water loops, and chilled water loops.  The new structure also includes a
system of checks to communicate to users if they are trying to assign a
certain feature that does not exist on a given HVAC system (for example,
cooling system parameters on a warm air furnace system).  It's
simplicity is meant to enable the easy implementation of different types
of systems (as opposed to a really detailed implementation of a few
systems).  Keep your eyes peeled for tomorrow's commit that should
result in components that you can test.