Fully Implemented ASHRAE Baseline Systems 0-8
@mostaphaRoudsari , @LelandCurtis  and @antonszilasi ,
I have been able to get through a clean implementation of ASHRAE
Baseline Systems 0-8 and I will polish off systems 9 and 10 tomorrow.
The important thing about this big cleanup effort is that all of the
parameters that you can assign on the airDetail, heatingDetail, and
coolingDetail components work.  If a certain system does not support the
assigning of a given parameter, you will get an error on the
OPenStudio_HVAC Systems component.  In this way, you finally can trust
that all of the changes you make in GH will actually affect the
simulation (or you'll get a warning if they won't).
At this point in time, there are just a couple of parameters assigned on
the zone level that do not yet affect the OpenStudio HVAC simulation and
these are 1) humidity control and 2) ventilation schedules (note that
ventilation loads will affect the simulation, though).  I'll put these
capabilities in shortly after implementing systems 9 and 10.
In the meantime, please give the new components a test drive and let me
know your thoughts.
Thanks, as always,
-Chris