Clean Implementation of All 10 ASHRAE Baseline Systems
@mostaphaRoudsari , @LelandCurtis , @BoLinAE , @antonszilasi ,
At this point, the only things left to implement for full OpenStudio
HVAC capabilities are humidity control and ventilation schedules, which
I will put in shortly.  And, once we agree that this current
implementation is stable, we should take out a lot of hard-coded stuff
that is not being used right now.