Humidification Working for System 7. VAV w/ Reheat
@mostaphaRoudsari , @LelandCurtis , and @BoLinAE ,
Let me know your thoughts on this implementation that I am doing for
humidity control on air-based systems.  For dehumidification, I just run
the cooling coil more intensely (which I know is done in practice a
lot).  However, for humidification, it seems the only gauranteed option
that OpenStudio has at the moment is to put an electric resistance
humidifier into the air loop.  Most of the time, I imagine that
humidification won't come into play but let me know if you think that
there is a better way of implementing this.  Generally, electric
resistance heating just makes me feel gross from an exergy matching
standpoint.  This may just have to hold for the time being, though.