Properly Added Humidity Control
It was giving some results with high cooling energy use because it was
poorly interacting with the economizer but this is now fixed.  Right
now, I am adding the humidity controller to the supply outlet node,
which works but is not really what EnergyPlus likes.  I cannot add it to
the node that E+ likes because of a bug in OpenStudio which causes Rhino
to crash if I try to .get() the right node.