Implemented First OS HVAC Beyond ASHRAE Baselines
@mostaphaRoudsari , @LelandCurtis , @BoLinAE , @antonszilasi ,
I am happy to announce the implementation of the first HVAC system
beyond the 10 ASHRAE baseline systems.  The new system is Fan Coil Units
+ DOAS but, more importantly, the means by which I have implemented it
makes it very easy to add system tpyes for chilled beams, radiant slabs,
and GSHPs, which will all come soon.

Importantly, I would like to hear any of your thoughts on how I have
implemented this.  I realized that all of the HVAC measures that have
been put out by NREL are using the same OSLib_HVAC.rb library of
functions in order to help add in new HVAC systems.  In this commit, I
have essentially added a python version of this ruby library into the
Export to OpenStudio component (although only a third of the functions
from the original library have been translated and I will only add more
functions as the demand for different system types arises).  While this
is a lot of additional code up front, I think it will save us a lot of
time in the long run as we implement more and more system types and try
to make airDetails, heatingDetails, and coolingDetals impact the system.
In other words, starting with an ASHRAE baseline system for each new
system and editing it seems like a big hassle over the long run compared
to this library method.  Please let me know if you agree and, if so,
whether the ExportToOpenStudio is the right place for this library (or
should it live in HB_HB?)