Automatically Split THERM Polygons with Holes Before Export
I realized that a lot of people want to be able to input a surface with
holes for their polygons but THERM does not like this since all polygons
must have a single boundary.  Splitting the geometry before export is
relatively easy and it is something that I now take care of for the user
(similar to how the E+ export takes care of meshing your curved surfaces
before export).