Switched out PopUp Window for Printed File
@MingboPeng ,
I'm really sorry but the popup window became annoying quickly.  This is
particularly because there is no code that automatically updates the
updateLog.md.  Nor is there any code that checks the new updateLog
against the old one.  I changed the component to print the updatLog file
from the component if you still want to add these features.  If not,
maybe we should just take out the updateLog.md