Added Dynamic Blinds To OpenStudio
This took a while but I have re-written all shade/blind materials made
by the 'Window Shade Generator' to be EPObjects.  I also made all
shadingControl EPObjects and reduced the parameter on the HBZones to be
just the name of these objects instead of a whole IDF string.  This
results in a much lighter IDF size, smaller GH memory use, and also
allowed me to easily implement dynamic shades in OpenStudio.
...Well. almost easily.  It turns out that OpenStudio has not yet
implemented any other shade control type except for
'OnIfHighSolarOnWindow'.  As a result, I had to write a workaround to
enable all of the useful shading control options in the IDF after it is
written by OpenStudio..
This should finish off the enabling of all non-WIP Honeybee features on
the OpenStudio component.    Welcome to the age of OpenStudio!!!!!!!