Fixed Up AutoGen of DDY File
Because OpenStudio requires Design Day Objects in order to run systems
correctly, we need a function that is able to deal with the case that
the user does not have a DDY file.  Here, I have added a function that
is able to generate DDY objects from the EPW.  The resulting DDY is very
close to those distributed by ASHRAE.  I am able to match the heating
sizing almost exactly.  The cooling sizing is off by ~2% because the Sky
model for radiation used by AHSRAE requires some information not
contained in the EPW.  Instead, I use ASHRAE's simpler ClearSky model at
the highest solar intensity.  As a result, the function here oversizes
the cooling system by small amount (~2% as mentioned previously).