Changed E+ Install to OS Install
Honeybee_Honeybee now checks for any installed OpenStudio versions and
gives a warning if it cannot find any.  The  OpenStudio C# library path
is copied to the sticky so that it can be used by the other components
(no need to check on the OS component).  Finally, the E+ component can
now run off of the EnergyPlus that is installed in the OpenStudio folder
so there is no need to have two installations of E+ on one's machine
(just install OpenStudio and you are set!)