Avoid Sorting HVAC Systems Based on Unique ID
I realized that the order (and corresponding number) of the system being
written into OpenStudio was set by the alphabetic sorting of the unique
ID's for the HVAC.  This was resulting in a loss of information about
which system corresponded to which zone.  I have re-written this to sort
the HVAC labels based on the order that HBzones are connected to the
OpenStudio component.  This makes it easier to know which system number
goes with each zone.