Distinguish Load from Energy in Thermal Results
In respect of the second law of thermodynamics, I am trying to make it
clear when the user is simply observing a thermal load or actual
fuel/electric energy.  Also, I only make the energy balance and total
thermal outputs appear when one has modeled an ideal air loads system.