Modified Honeybee_hive
Honeybee hive used to create a copy of each object twice and give a
unique id to each item. The current implementation cleans up the hive
each time that there is a new set of geometries from the same component.
Unfortunately, the last iteration will stay  in the memory as we can't
overwrite/extend OnDelete method of the component. This can be addressed
in Rhino 6.

One other major change is that now all the Set EP properties components
accept lists for zones and inputs. This makes the component more
flexible and faster and also helps the issues with the memory. The only
drawback is that one won't be able to use a list input for specifying
constructions for different orientations. This can be implemented in
Tree Access mode.
cc: @chriswmackey