Added Support for OpenStudio 2.1
everything seems to be working well with the new version.  The biggest
change is that we had to add in a workaround because OpenStudio is no
longer dsitributed with the files needed to run EnergyPlus from batch
files.  HB_HB will download these files from the github if it detects
2.1 without these files.