Cleaned Up Schedule Components
These are finally all working as originally intended, making it easy for
people to build schedules without having to resort to the CSV.  People
can also easily assign holidays and change the start day of the week for
both the simulation and the displayed "schedule values." There's only
one thing left that I want to add and that is seasonal schedules.  Then
I can officially say that there's nothing tha cannot be created with the
Honeybee schedule components.