Fixed Bug Because THERM is Weird
Man, this program has so many odd bugs.  I assume this is half because
THERM is young and still under development.  I feel that the other half
is that it's closed source, though.  Soneone would have identified and
fixed a bug like this already if it were open source.  I hope the LBNL
team gets there soon.