Major Update to Honeybee HVAC
This includes the ability to model central GSHP systems with hydronic distribution (like radiant systems with water heated/cooled by a GSHP).  It also includes a new convention for specifying ground sourced systems, which affects the existing GSHP templates and essentially converts them into VRF and WSHP templates with the option of being linked ot a ground loop.  Lastly, this commit includes the ability to set the capacity of the chiller/boiler/heat pump.  This was partially done to get around a bug in E+ with central heat pump systems but this also has other uses.