fix(openstudio): Correct OpenStudio version in the event that ladybug_tools folder is used
Since @mingbopeng isn't writing the version into the openstudio directory into the exe installer, there's no way to know the version for sure. So I'm just hard-setting it to version 3.0 and let's just hope that we never need to release the legacy executable installer with a later version of Openstudio.